Made login screen responsive
